feat: Enhance validation and error handling
- Creation of DataValidationException.java: Added a custom exception class for better validation error handling.
- Updates to CourseRegistrationDTO.java, TopicRegistrationDTO.java, and UserRegistrationDTO.java: Adjusted DTO classes to return specific validation messages.
- Changes to TopicRepository.java: Updated to assist with data validation according to business rules.
- Modifications in TopicService.java: Implemented error handling and validations in the topic service to ensure data integrity and proper exception handling.
- Creation of topic validators: TopicAuthorExistenceValidator.java, TopicCommonValidator.java, TopicCourseExistenceValidator.java, TopicRegistrationValidator.java, TopicUniquenessValidator.java, TopicExistenceValidator.java, TopicUpdateUniquenessValidator.java, TopicUpdateValidator.java.
- Creation of ErrorHandler.java: Improved error handling to manage new exceptions and provide more informative error messages.
- Updated configurations in application.properties: Adjusted properties to reflect new validation and error handling settings.
- Creation of validationMessages.properties: Added custom validation messages.

+++ src/main/java/forum/hub/api/domain/topic/TopicService.java
@@ -1,8 +1,12 @@
 package forum.hub.api.domain.topic;
 
+import forum.hub.api.domain.DataValidationException;
 import forum.hub.api.domain.course.CourseRepository;
 import forum.hub.api.domain.topic.dto.TopicDetailsDTO;
 import forum.hub.api.domain.topic.dto.TopicRegistrationDTO;
+import forum.hub.api.domain.topic.validations.general.TopicCommonValidator;
+import forum.hub.api.domain.topic.validations.registration.TopicRegistrationValidator;
+import forum.hub.api.domain.topic.validations.update.TopicUpdateValidator;
 import forum.hub.api.domain.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
@@ -10,6 +14,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Service
 public class TopicService {
@@ -19,8 +24,17 @@ public class TopicService {
     private UserRepository userRepository;
     @Autowired
     private CourseRepository courseRepository;
+    @Autowired
+    private List<TopicCommonValidator> topicCommonValidators;
+    @Autowired
+    private List<TopicRegistrationValidator> topicRegistrationValidators;
+    @Autowired
+    private List<TopicUpdateValidator> topicUpdateValidators;
 
     public TopicDetailsDTO register(TopicRegistrationDTO data) {
+        topicCommonValidators.forEach(v -> v.validate(data));
+        topicRegistrationValidators.forEach(v -> v.validate(data));
+
         var author = userRepository.getReferenceById(data.authorId());
         var course = courseRepository.getReferenceById(data.courseId());
         var topic = new Topic(null, data.title(), data.message(), LocalDateTime.now(), Status.ABERTO, author, course);
@@ -41,6 +55,9 @@ public class TopicService {
     }
 
     public TopicDetailsDTO update(TopicRegistrationDTO data, Long id) {
+        topicCommonValidators.forEach(v -> v.validate(data));
+        topicUpdateValidators.forEach(v -> v.validate(data, id));
+
         var topic = topicRepository.getReferenceById(id);
         var user = userRepository.getReferenceById(data.authorId());
         var course = courseRepository.getReferenceById(data.courseId());
@@ -50,6 +67,12 @@ public class TopicService {
     }
 
     public void delete(Long id) {
+
+        var topicExists = topicRepository.existsById(id);
+
+        if (!topicExists){
+            throw new DataValidationException("Tópico não encontrado");
+        }
         topicRepository.deleteById(id);
     }
 }
